Diving Decoder tests in several files
